nevo-backend-framework-main/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "nevo_framework"
dynamic = ["version"]
requires-python = ">=3.11"
authors = [
    {name="David Blumenthal-Barby", email="david.blumenthal-barby@diconium.com"},
    {name="Neil Sinclair", email="neil.sinclair@diconium.com"},
    {name="Marshall Mykietyshyn", email="marshall.mykietyshyn@diconium.com"},
]
description = "This package works as the backend framework for Nevo sales apps."
readme = {file = "README.md", content-type = "text/markdown"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "distro==1.9.0",
    "fastapi==0.115.14",
    "frozenlist==1.7.0",
    "google-auth==2.40.3",
    "google-genai==1.24.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "multidict==6.6.3",
    "numpy==2.3.1",
    "openai==1.93.0",
    "propcache==0.3.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyaudio==0.2.14",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "tenacity==8.5.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn[standard]==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "yarl==1.20.1"
]

[project.urls]
Homepage = "https://github.com/diconium/nevo-backend-framework"
Issues = "https://github.com/diconium/nevo-backend-framework/issues"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"